+++ src/index.tsx
@@ -1,8 +1,15 @@
 import { NativeModules } from 'react-native';
 
 type WeFitterHealthKitType = {
-  connect(token: string): void;
+  configure(config: {
+    token: string;
+    url?: string;
+    startDate?: string;
+  }): Promise<void>;
+  connect(): Promise<void>;
   disconnect(): void;
+  canConnectToHealthData(callback: (supported: boolean) => void): void;
+  getConnectedProfileId(callback: (id: string) => void): void;
   getStatus(): void;
 };