+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## 1.0.0
+
+Updated WeFitter API version to 1.3
+
+Enabled background delivery for all supported datatypes
+
+Added `configure`, `canConnectToHealthData` and `getConnectedProfileId`
+
+Added `not-configured` status event
+
+BREAKING CHANGE:
+
+Removed param from `connect`
+
 ## 0.2.1
 
 Fixed bug preventing upload of daily summaries